improve the common module import
